test: share installHookDom via rendererTestHarness
The fake-DOM setup for mounting hook harnesses was copied in
liveTranscriptHoldLifecycle.test.js and gpuBannerAvailability.test.js.
Hoist the superset (with rAF stubs) into test/lib/rendererTestHarness.js
and import it from both.
