+++ test/helpers/liveTranscriptHoldLifecycle.test.js
@@ -2,66 +2,13 @@ const test = require("node:test");
 const assert = require("node:assert/strict");
 const React = require("react");
 const { createRoot } = require("react-dom/client");
-const { createRendererServer, installBrowserGlobals } = require("../lib/rendererTestHarness");
+const {
+  createRendererServer,
+  installBrowserGlobals,
+  installHookDom,
+} = require("../lib/rendererTestHarness");
 
 const FINAL_HIDE_MS = 4000;
-
-function installHookDom(t) {
-  const originalDocument = globalThis.document;
-  const originalActEnvironment = globalThis.IS_REACT_ACT_ENVIRONMENT;
-  const originalRequestAnimationFrame = globalThis.requestAnimationFrame;
-  const originalCancelAnimationFrame = globalThis.cancelAnimationFrame;
-  const noop = () => {};
-
-  class Element {}
-  class HTMLElement extends Element {}
-  class HTMLIFrameElement extends HTMLElement {}
-
-  const document = {
-    nodeType: 9,
-    activeElement: null,
-    addEventListener: noop,
-    removeEventListener: noop,
-  };
-  const container = {
-    nodeType: 1,
-    nodeName: "DIV",
-    tagName: "DIV",
-    namespaceURI: "http://www.w3.org/1999/xhtml",
-    ownerDocument: document,
-    addEventListener: noop,
-    removeEventListener: noop,
-    appendChild: noop,
-    removeChild: noop,
-    insertBefore: noop,
-  };
-  Object.assign(globalThis.window, {
-    Element,
-    HTMLElement,
-    HTMLIFrameElement,
-    document,
-    getSelection: () => null,
-  });
-  document.defaultView = globalThis.window;
-  document.documentElement = container;
-  globalThis.document = document;
-  globalThis.IS_REACT_ACT_ENVIRONMENT = true;
-  globalThis.requestAnimationFrame = () => 1;
-  globalThis.cancelAnimationFrame = noop;
-
-  t.after(() => {
-    if (originalDocument === undefined) delete globalThis.document;
-    else globalThis.document = originalDocument;
-    if (originalActEnvironment === undefined) delete globalThis.IS_REACT_ACT_ENVIRONMENT;
-    else globalThis.IS_REACT_ACT_ENVIRONMENT = originalActEnvironment;
-    if (originalRequestAnimationFrame === undefined) delete globalThis.requestAnimationFrame;
-    else globalThis.requestAnimationFrame = originalRequestAnimationFrame;
-    if (originalCancelAnimationFrame === undefined) delete globalThis.cancelAnimationFrame;
-    else globalThis.cancelAnimationFrame = originalCancelAnimationFrame;
-  });
-
-  return container;
-}
 
 function capturePanelTimers(t) {
   const originalSetTimeout = globalThis.setTimeout;
