+++ test/lib/rendererTestHarness.js
@@ -35,6 +35,66 @@ function installBrowserGlobals(t, { initialStorage = {}, window: windowProps = {
   return { window: globalThis.window, storage };
 }
 
+// Minimal fake DOM for mounting hook harnesses with react-dom's createRoot:
+// just enough node structure for React to attach a root — no layout, no real
+// events. Call installBrowserGlobals first; this assigns onto globalThis.window.
+function installHookDom(t) {
+  const originalDocument = globalThis.document;
+  const originalActEnvironment = globalThis.IS_REACT_ACT_ENVIRONMENT;
+  const originalRequestAnimationFrame = globalThis.requestAnimationFrame;
+  const originalCancelAnimationFrame = globalThis.cancelAnimationFrame;
+  const noop = () => {};
+
+  class Element {}
+  class HTMLElement extends Element {}
+  class HTMLIFrameElement extends HTMLElement {}
+
+  const document = {
+    nodeType: 9,
+    activeElement: null,
+    addEventListener: noop,
+    removeEventListener: noop,
+  };
+  const container = {
+    nodeType: 1,
+    nodeName: "DIV",
+    tagName: "DIV",
+    namespaceURI: "http://www.w3.org/1999/xhtml",
+    ownerDocument: document,
+    addEventListener: noop,
+    removeEventListener: noop,
+    appendChild: noop,
+    removeChild: noop,
+    insertBefore: noop,
+  };
+  Object.assign(globalThis.window, {
+    Element,
+    HTMLElement,
+    HTMLIFrameElement,
+    document,
+    getSelection: () => null,
+  });
+  document.defaultView = globalThis.window;
+  document.documentElement = container;
+  globalThis.document = document;
+  globalThis.IS_REACT_ACT_ENVIRONMENT = true;
+  globalThis.requestAnimationFrame = () => 1;
+  globalThis.cancelAnimationFrame = noop;
+
+  t.after(() => {
+    if (originalDocument === undefined) delete globalThis.document;
+    else globalThis.document = originalDocument;
+    if (originalActEnvironment === undefined) delete globalThis.IS_REACT_ACT_ENVIRONMENT;
+    else globalThis.IS_REACT_ACT_ENVIRONMENT = originalActEnvironment;
+    if (originalRequestAnimationFrame === undefined) delete globalThis.requestAnimationFrame;
+    else globalThis.requestAnimationFrame = originalRequestAnimationFrame;
+    if (originalCancelAnimationFrame === undefined) delete globalThis.cancelAnimationFrame;
+    else globalThis.cancelAnimationFrame = originalCancelAnimationFrame;
+  });
+
+  return container;
+}
+
 // mockModules maps an import-path suffix (e.g. "/utils/logger") to the ESM
 // source served in its place.
 async function createRendererServer(
@@ -76,4 +136,4 @@ async function createRendererServer(
   return vite;
 }
 
-module.exports = { createRendererServer, installBrowserGlobals };
+module.exports = { createRendererServer, installBrowserGlobals, installHookDom };
